backend/utils/supabase_client.py
import os
from supabase import create_client, Client
from dotenv import load_dotenv

load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_ANON_KEY")   # <<< AJUSTADO

if not SUPABASE_URL or not SUPABASE_KEY:
    raise Exception("❌ SUPABASE_URL ou SUPABASE_ANON_KEY não configuradas!")

supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)
