backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose
passlib[bcrypt]
python-dotenv
supabase
httpx

